moving and renaming cookie-collector
